Proteger rutas con gates

+++ app/Http/Controllers/RoleController.php
@@ -2,19 +2,24 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
 use Illuminate\Http\Request;
 use App\Base\Models\Permission;
 use App\Base\Models\Role;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\View\View;
 
 class RoleController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Application|Factory|View
      */
     public function index()
     {
+        $this->authorize('have-access', 'roles.index');
         $roles = Role::latest()->paginate(10);
         return view('roles.index', compact('roles'));
     }
@@ -26,6 +31,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        $this->authorize('have-access', 'roles.create');
         $permissions = Permission::get();
         return view('roles.create', compact('permissions'));
     }
@@ -37,7 +43,8 @@ class RoleController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {    
+    {
+        $this->authorize('have-access', 'roles.create');
        $request->validate([
             'name'        => 'required|max:30|unique:roles,name',
             'slug'        => 'required|unique:roles,slug',
@@ -50,7 +57,7 @@ class RoleController extends Controller
         if ($request->input('permissions')) {
             $role->permissions()->sync($request->input('permissions'));
         }
-        
+
         return redirect()->route('roles.index')->with('status', __('Role saved successfully'));
     }
 
@@ -62,6 +69,7 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        $this->authorize('have-access', 'roles.show');
         $permissionsRole = [];
 
         foreach ($role->permissions as $permission) {
@@ -80,6 +88,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('have-access', 'roles.edit');
         $permissionsRole = [];
 
         foreach ($role->permissions as $permission) {
@@ -98,22 +107,25 @@ class RoleController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Role $role)
-    {   
+    {
+        $this->authorize('have-access', 'roles.edit');
         $request->validate([
              'name'        => 'required|max:30|unique:roles,name,'.$role->id,
              'slug'        => 'required|unique:roles,slug,'.$role->id,
              'description' => 'max:500|min:3',
              'full-access' => 'required|in:yes,not',
          ]);
- 
+
          $role->update($request->all());
- 
+
          if ($request->input('permissions')) {
              $role->permissions()->sync($request->input('permissions'));
+         } else {
+             $role->permissions()->sync([]);
          }
-         
+
          return redirect()->route('roles.index')->with('status', __('Role updated successfully'));
- 
+
     }
 
     /**
@@ -124,6 +136,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        $this->authorize('have-access', 'roles.destroy');
         $role->delete();
         return redirect()->route('roles.index')->with('status', __('Role deleted successfully'));
     }
